+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Apache Arrow Flight-based query interface now transfers data in batches.
 - Reverted to the key-value configuration file format and removed PureConfig.
   The data transfer module is configured with the `modelardb.transfer` setting.
+- The value of `modelardb.dynamic_split_fraction` is now used directly as set.
 
 ### Deprecated
 ### Removed